moved layout type components for a separeted folder and updated the paths for each
