+++ src/components/layout/Hero.js
@@ -1,13 +1,13 @@
 // Components
-import Button from "../components/Button"
-import Image from "../components/Image"
+import Button from "../../components/Button"
+import Image from "../../components/Image"
 
 // Assets
-import heroImage from "../assets/Hero-Image.png"
-import sectionImage from "../assets/Section-Image-2.png"
+import heroImage from "../../assets/Hero-Image.png"
+import sectionImage from "../../assets/Section-Image-2.png"
 
 // Stylesheets
-import "../styles/Hero.css"
+import "../../styles/Hero.css"
 
 function Hero() {
   return (
