+++ src/components/layout/Section.js
@@ -1,9 +1,9 @@
 // Assets
-import sectionImage2 from "../assets/Section-Image-2.png"
-import sectionImage3 from "../assets/Section-Image-3.png"
+import sectionImage2 from "../../assets/Section-Image-2.png"
+import sectionImage3 from "../../assets/Section-Image-3.png"
 
 // Stylesheets
-import "../styles/Section.css"
+import "../../styles/Section.css"
 
 function Section({section}) {
   if (section == "participate") {
